video 149 - nginx proxy for websocket
Also fixed some typos and got the code a bit
better aligned with the checkpoint code. There is
still an error in it, somewhere around the redis
worker, as logging into the postgres container I
can see the values which I've added, at least some
of them anyway. This probably bears further investigation.